# cmd/api/docs/swagger.yaml
basePath: /v1
definitions:
  cerrors.IError:
    type: object
...
  database.SubRating:
    properties:
      count:
        type: integer
      users:
        items:
          properties:
            avatarURL:
              type: string
            login:
              type: string
          type: object
        type: array
    type: object
  docstring.Typedef:
    properties:
      args:
        items:
          $ref: '#/definitions/docstring.TypedefArg'
        type: array
      name:
        type: string
      type:
        type: string
    type: object
  docstring.TypedefArg:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  gin.H:
    additionalProperties: true
    type: object
...
  handlers.Contract:
    properties:
      address:
        type: string
      alias:
        type: string
      annotations:
        items:
          type: string
        type: array
      balance:
        type: integer
      delegate:
        type: string
      delegate_alias:
        type: string
      entrypoints:
        items:
          type: string
        type: array
      fail_strings:
        items:
          type: string
        type: array
      fingerprint:
        $ref: '#/definitions/models.Fingerprint'
        type: object
      found_by:
        type: string
      hardcoded:
        items:
          type: string
        type: array
      hash:
        type: string
      id:
        type: string
      language:
        type: string
      last_action:
        $ref: '#/definitions/models.BCDTime'
        type: object
      level:
        type: integer
      manager:
        type: string
      migrations_count:
        type: integer
      network:
        type: string
      profile:
        $ref: '#/definitions/handlers.ProfileInfo'
        type: object
      project_id:
        type: string
      slug:
        type: string
      tags:
        items:
          type: string
        type: array
      timestamp:
        type: string
      total_withdrawn:
        type: integer
      tx_count:
        type: integer
    type: object
  handlers.EntrypointSchema:
    properties:
      bin_path:
        type: string
      default_model:
        $ref: '#/definitions/jsonschema.DefaultModel'
        type: object
      name:
        type: string
      schema:
        $ref: '#/definitions/jsonschema.Schema'
        type: object
      typedef:
        items:
          $ref: '#/definitions/docstring.Typedef'
        type: array
    type: object
  handlers.Error:
    properties:
      message:
        example: text
        type: string
    type: object
...
  handlers.NetworkStats:
    properties:
      contracts_count:
        example: 10
        type: integer
      operations_count:
        example: 100
        type: integer
      protocols:
        items:
          $ref: '#/definitions/models.Protocol'
        type: array
    type: object
  handlers.Operation:
    properties:
      amount:
        type: integer
      balance:
        type: integer
      balance_updates:
        items:
          $ref: '#/definitions/models.BalanceUpdate'
        type: array
      burned:
        type: integer
      counter:
        type: integer
      delegate:
        type: string
      destination:
        type: string
      destination_alias:
        type: string
      entrypoint:
        type: string
      errors:
        items:
          $ref: '#/definitions/cerrors.IError'
        type: array
      fee:
        type: integer
      gas_limit:
        type: integer
      hash:
        type: string
      id:
        type: string
      internal:
        type: boolean
      kind:
        type: string
      level:
        type: integer
      manager_pubkey:
        type: string
      mempool:
        type: boolean
      network:
        type: string
      parameters:
        type: object
      protocol:
        type: string
      public_key:
        type: string
      result:
        $ref: '#/definitions/models.OperationResult'
        type: object
      source:
        type: string
      source_alias:
        type: string
      status:
        type: string
      storage_diff:
        type: object
      storage_limit:
        type: integer
      timestamp:
        type: string
    type: object
...
  handlers.ProfileInfo:
    properties:
      subscribed:
        type: boolean
    type: object
...
  jsonschema.DefaultModel:
    additionalProperties: true
    type: object
  jsonschema.Schema:
    additionalProperties: true
    type: object
  models.BCDTime:
    type: object
  models.BalanceUpdate:
    properties:
      category:
        type: string
      change:
        type: integer
      contract:
        type: string
      cycle:
        type: integer
      delegate:
        type: string
      kind:
        type: string
    type: object
  models.Block:
    properties:
      chain_id:
        example: NetXdQprcVkpaWU
        type: string
      hash:
        example: BLyAEwaXShJuZasvUezHUfLqzZ48V8XrPvXF2wRaH15tmzEpsHT
        type: string
      level:
        example: 24
        type: integer
      network:
        example: mainnet
        type: string
      predecessor:
        example: BMWVEwEYw9m5iaHzqxDfkPzZTV4rhkSouRh3DkVMVGkxZ3EVaNs
        type: string
      protocol:
        example: PtCJ7pwoxe8JasnHY8YonnLYjcVHmhiARPJvqcC6VfHT5s8k8sY
        type: string
      timestamp:
        example: "2018-06-30T18:05:27Z"
        type: string
    type: object
...
  models.Fingerprint:
    properties:
      code:
        type: string
      parameter:
        type: string
      storage:
        type: string
    type: object
  models.OperationResult:
    properties:
      allocated_destination_contract:
        type: boolean
      balance_updates:
        items:
          $ref: '#/definitions/models.BalanceUpdate'
        type: array
      consumed_gas:
        type: integer
      paid_storage_size_diff:
        type: integer
      storage_size:
        type: integer
    type: object
  models.Protocol:
    properties:
      alias:
        example: Carthage
        type: string
      end_level:
        example: 0
        type: integer
      hash:
        example: PsCARTHAGazKbHtnKfLzQg3kms52kSRpgnDY982a9oYsSXRLQEb
        type: string
      network:
        example: mainnet
        type: string
      start_level:
        example: 851969
        type: integer
      sym_link:
        example: babylon
        type: string
    type: object
  newmiguel.Node:
    properties:
      children:
        items:
          $ref: '#/definitions/newmiguel.Node'
        type: array
      diff_type:
        type: string
      from:
        type: object
      name:
        type: string
      prim:
        type: string
      type:
        type: string
      value:
        type: object
    type: object
host: api.better-call.dev
info:
  contact: {}
  description: This is API description for Better Call Dev service.
  license: {}
  title: Better Call Dev API
  version: "1.0"
paths:
  /contract/{network}/{address}:
    get:
      consumes:
      - application/json
      description: Get full contract info
      operationId: get-contract
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Contract'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get contract info
      tags:
      - contract
...
  /contract/{network}/{address}/entrypoints:
    get:
      consumes:
      - application/json
      description: Get contract entrypoints
      operationId: get-contract-entrypoints
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.EntrypointSchema'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get contract entrypoints
      tags:
      - contract
...
  /contract/{network}/{address}/entrypoints/trace:
    post:
      consumes:
      - application/json
      description: Execute entrypoint with passed arguments
      operationId: get-contract-entrypoints-run
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      - description: Entrypoint binary path
        in: body
        name: bin_path
        required: true
        schema:
          type: string
      - description: Entrypoint`s arguments data
        in: body
        name: data
        required: true
        schema:
          type: object
      - description: Operation amount
        in: body
        name: amount
        schema:
          type: integer
      - description: Operation gas limit
        in: body
        name: gas_limit
        schema:
          type: integer
      - description: Operation sender
        in: body
        name: sender
        schema:
          type: string
      - description: Operation source
        in: body
        name: source
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.Operation'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Execute entrypoint with passed arguments
      tags:
      - contract
...
  /contract/{network}/{address}/rating:
    get:
      consumes:
      - application/json
      description: Get contract rating
      operationId: get-contract-rating
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.SubRating'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get contract rating
      tags:
      - contract
  /contract/{network}/{address}/raw_storage:
    get:
      consumes:
      - application/json
      description: Get contract raw storage
      operationId: get-contract-storage-raw
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get contract raw storage
      tags:
      - contract
  /contract/{network}/{address}/rich_storage:
    get:
      consumes:
      - application/json
      description: Get contract rich storage
      operationId: get-contract-storage-rich
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get contract rich storage
      tags:
      - contract
...
  /contract/{network}/{address}/storage:
    get:
      consumes:
      - application/json
      description: Get contract storage
      operationId: get-contract-storage
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: KT address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/newmiguel.Node'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get contract storage
      tags:
      - contract
...
  /stats:
    get:
      consumes:
      - application/json
      description: get indexer states for all networks
      operationId: get-stats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Block'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Show indexer stats
      tags:
      - statistics
  /stats/{network}:
    get:
      consumes:
      - application/json
      description: Get detailed statistics for network
      operationId: get-network-stats
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.NetworkStats'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Network statistics
      tags:
      - statistics
  /stats/{network}/series:
    get:
      consumes:
      - application/json
      description: Get count series data for network
      operationId: get-network-stats
      parameters:
      - description: Network
        in: path
        name: network
        required: true
        type: string
      - description: One of index name (contract, operation)
        enum:
        - contract
        - operation
        in: query
        name: index
        required: true
        type: string
      - description: One of period (year, month, week, day)
        enum:
        - year
        - month
        - week
        - day
        in: query
        name: period
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.Error'
      summary: Get network series
      tags:
      - statistics
...
swagger: "2.0"